# functions/lib/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "FeedSolve API",
        "version": "1.0.0",
        "description": "REST API for feedback management and analytics",
        "contact": {
            "name": "FeedSolve Support",
            "url": "https://feedsolve.com/support"
        }
    },
    "servers": [
        {
            "url": "https://api.feedsolve.com",
            "description": "Production API"
        },
        {
            "url": "http://localhost:5001",
            "description": "Local development"
        }
    ],
    "security": [
        {
            "BearerAuth": []
        }
    ],
    "paths": {
...
        "/api/boards": {
            "post": {
                "summary": "Create a new board",
                "tags": ["Boards"],
                "operationId": "createBoard",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BoardInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Board created successfully"
                    },
                    "400": {
                        "description": "Invalid input"
                    }
                }
            }
        },
        "/api/boards/{id}": {
            "get": {
                "summary": "Get board by ID",
                "tags": ["Boards"],
                "operationId": "getBoard",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Board found"
                    },
                    "404": {
                        "description": "Board not found"
                    }
                }
            },
            "patch": {
                "summary": "Update board",
                "tags": ["Boards"],
                "operationId": "updateBoard",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BoardInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Board updated"
                    },
                    "404": {
                        "description": "Board not found"
                    }
                }
            },
            "delete": {
                "summary": "Delete board",
                "tags": ["Boards"],
                "operationId": "deleteBoard",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Board deleted"
                    },
                    "404": {
                        "description": "Board not found"
                    }
                }
            }
        },
...
        "/api/boards/{id}/stats": {
            "get": {
                "summary": "Get board statistics",
                "tags": ["Analytics"],
                "operationId": "getBoardStats",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Board statistics"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BoardInput": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "description": {
                        "type": "string"
                    },
                    "categories": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "isAnonymousAllowed": {
                        "type": "boolean"
                    }
                },
                "required": ["name"]
            },
...
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "token",
                "description": "API Key in format: Bearer fsk_xxxxx"
            }
        }
    }
}